refactor: moved the mysql provider support into separate directory
